予約の有効無効を録画モードから分離( xtne6f@6c9ce5f )

+++ EpgTimer/EpgTimer/Menu/CmdExeAutoAdd.cs
@@ -13,7 +13,7 @@ namespace EpgTimer
         public CmdExeAutoAdd(UIElement owner) : base(owner) { }
         protected override void mc_ChangeKeyEnabled(object sender, ExecutedRoutedEventArgs e)
         {
-            IsCommandExecuted = MenuUtil.AutoAddChangeKeyEnabled(dataList, (byte)CmdExeUtil.ReadIdData(e, 0, 1) == 0);
+            IsCommandExecuted = MenuUtil.AutoAddChangeKeyEnabled(dataList, CmdExeUtil.ReadIdData(e, 0, 1) == 0);
         }
         protected override void mc_ChangeOnOffKeyEnabled(object sender, ExecutedRoutedEventArgs e)
         {
